Rename mulliken module to partition module
I intend to add some sort of bond order analysis method here
